requirements.txt
openai>=1.0.0
stability-sdk>=0.8.0
pillow>=9.0.0
numpy>=1.21.0
torch>=1.12.0
opencv-python>=4.5.0
matplotlib>=3.5.0
IPython>=8.0.0
requests>=2.28.0 